requirements.txt
slack_sdk>=3.0
slackeventsapi>=2.1.0
Flask>=1.1.2
Pandas
datetime
openpyxl
request